fix: edit catalog styling, description sheet
