+++ package.json
@@ -37,7 +37,6 @@
     "next": "^14.2.15",
     "next-themes": "^0.4.3",
     "react": "^18",
-    "react-aria-components": "^1.3.3",
     "react-dom": "^18",
     "react-icons": "^5.3.0",
     "react-slick": "^0.30.2",
