+++ src/app/(views)/catalogs/[catalogId]/edit/catalog-form.tsx
@@ -3,14 +3,12 @@
 import { useEffect, useState } from "react";
 import { z } from "zod";
 
-import { Input, Label } from "~/components/custom/Field";
-import Spinner from "~/components/custom/spinner.client";
+import Spinner from "~/components/custom/spinner";
 import { Button } from "~/components/shadcn/button";
+import { Input } from "~/components/shadcn/input";
+import { Label } from "~/components/shadcn/label";
 import { useToast } from "~/hooks/use-toast";
 import fetchApi from "~/utils/fetch";
-
-const YouTubeVideoLinkRegex =
-  /^.*(?:(?:youtu\.be\/|v\/|vi\/|u\/\w\/|embed\/|shorts\/)|(?:(?:watch)?\?v(?:i)?=|\&v(?:i)?=))([^#\&\?]*).*/;
 
 const CatalogSchema = z.object({
   title: z
@@ -22,11 +20,6 @@ const CatalogSchema = z.object({
     .min(8, { message: "Description must be at least 8 characters long." })
     .max(64, { message: "Description must be at most 64 characters long." }),
 });
-
-type VideoLink = {
-  link: string;
-  error: string;
-};
 
 type CatalogMetadata = z.infer<typeof CatalogSchema>;
 
@@ -54,11 +47,6 @@ export default function CatalogForm({
       title: "",
       description: "",
     });
-
-  const [videoLink, setVideoLink] = useState<VideoLink>({
-    link: "",
-    error: "",
-  });
 
   const [isSubmitting, setIsSubmitting] = useState<boolean>(false);
 
@@ -101,74 +89,6 @@ export default function CatalogForm({
     setLocalChannel([]);
   };
 
-  const validateVideoLink = async () => {
-    const found = videoLink.link.match(YouTubeVideoLinkRegex);
-    let videoId = "";
-    if (found?.length) {
-      videoId = found[1];
-    }
-
-    if (videoId) {
-      const result = await fetchApi(`/youtube/videoId?videoId=${videoId}`);
-
-      if (!result.success) {
-        toast({ title: result.message });
-        return;
-      }
-      const videoData = result.data.items[0].snippet;
-      const channelId = videoData.channelId;
-      const channelTitle = videoData.channelTitle;
-
-      const alreadyExists = localChannel.find(
-        (item: any) => item.id === channelId
-      );
-
-      const alreadySaved = savedChannels?.find(
-        (channel: any) => channel.id === channelId
-      );
-
-      if (!alreadyExists && !alreadySaved) {
-        setLocalChannel((prev: any) => [
-          ...prev,
-          { id: channelId, title: channelTitle },
-        ]);
-        toast({ title: `${channelTitle}'s channel added to the list.` });
-      } else if (alreadyExists) {
-        toast({
-          title: `${channelTitle}'s channel already added to the list.`,
-        });
-      } else if (alreadySaved) {
-        toast({ title: `${channelTitle}'s channel already saved.` });
-      }
-
-      setVideoLink({
-        link: "",
-        error: "",
-      });
-    }
-  };
-
-  const handleVideoLink = (e: any) => {
-    setVideoLink((prev) => ({
-      ...prev,
-      link: e.target.value,
-    }));
-
-    if (!YouTubeVideoLinkRegex.test(e.target.value)) {
-      setVideoLink((prev) => ({
-        ...prev,
-        error: "Invalid YouTube video link.",
-      }));
-
-      return;
-    } else {
-      setVideoLink((prev) => ({
-        ...prev,
-        error: "",
-      }));
-    }
-  };
-
   const handleMetaUpdate = (e: any) => {
     setCatalogMetadata((prev) => ({
       ...prev,
@@ -201,55 +121,39 @@ export default function CatalogForm({
     <>
       <div className="space-y-4 mt-5">
         <div className="flex flex-col gap-1">
-          <Label htmlFor="title">Title</Label>
-          <Input
-            value={catalogMetadata.title}
-            onChange={handleMetaUpdate}
-            name="title"
-            id="title"
-          />
-          {catalogMetadataError.title ? (
-            <p className="text-sm text-red-800">{catalogMetadataError.title}</p>
-          ) : null}
-        </div>
-        <div className="flex flex-col gap-1">
-          <Label htmlFor="description">Description</Label>
-          <Input
-            value={catalogMetadata.description}
-            onChange={handleMetaUpdate}
-            name="description"
-            id="description"
-          />
-          {catalogMetadataError.description ? (
-            <p className="text-sm text-red-800">
-              {catalogMetadataError.description}
-            </p>
-          ) : null}
-        </div>
-
-        <div className="flex flex-col gap-1">
-          <Label htmlFor="add-video">Add channel from video</Label>
-          <div className="flex gap-2 items-center">
+          <div className="space-y-2">
+            <Label htmlFor="title">Title</Label>
             <Input
-              id="add-video"
-              value={videoLink.link}
-              onChange={handleVideoLink}
+              id="title"
+              value={catalogMetadata.title}
+              name="title"
+              onChange={handleMetaUpdate}
             />
-            <Button
-              disabled={Boolean(videoLink.error || !videoLink.link)}
-              onClick={validateVideoLink}
-            >
-              Validate
-            </Button>{" "}
+            {catalogMetadataError.title ? (
+              <p className="text-sm text-[hsl(var(--primary))]">
+                {catalogMetadataError.title}
+              </p>
+            ) : null}
           </div>
-          {videoLink.error ? (
-            <p className="text-sm text-red-800">{videoLink.error}</p>
-          ) : null}
+
+          <div className="space-y-2">
+            <Label htmlFor="description">Description</Label>
+            <Input
+              id="description"
+              value={catalogMetadata.description}
+              name="description"
+              onChange={handleMetaUpdate}
+            />
+            {catalogMetadataError.description ? (
+              <p className="text-sm text-[hsl(var(--primary))]">
+                {catalogMetadataError.description}
+              </p>
+            ) : null}
+          </div>
         </div>
       </div>
 
-      <div className="flex gap-2 mt-5">
-        {/* TODO: Maybe update "Not Saved" channels and title/description separately */}
+      <div className="flex gap-2 mt-5 justify-end">
         <Button
           disabled={Boolean(
             (!localChannel.length &&
@@ -260,7 +164,7 @@ export default function CatalogForm({
           onClick={handleSubmit}
         >
           {isSubmitting ? <Spinner className="size-4" /> : null}
-          Submit
+          Apply
         </Button>
       </div>
     </>
