+++ src/app/(views)/explore/page.tsx
@@ -1,10 +1,9 @@
-import CatalogExplore from "./catalog-explore";
 import { CatalogExplorer } from "./catalog-explorer";
 
 export default async function ExplorePage() {
   return (
     <div className="p-3">
-      <h1 className="text-2xl font-semibold tracking-tight">Explore Catalog</h1>
+      <h1 className="text-2xl font-semibold tracking-tight">Explore</h1>
       <CatalogExplorer />
     </div>
   );
