+++ src/components/custom/icons.tsx
@@ -2,16 +2,19 @@ import { BiLogoGmail as GmailIcon } from "react-icons/bi";
 import { FaLink as LinkIcon } from "react-icons/fa6";
 import { FiGithub as GithubIcon } from "react-icons/fi";
 import { HiOutlinePencilAlt as EditIcon } from "react-icons/hi";
+import { IoIosCopy as CopyIcon } from "react-icons/io";
+import { IoShareSocial as ShareIcon } from "react-icons/io5";
 import { MdOutlineLogout as LogoutIcon } from "react-icons/md";
 import { MdDeleteForever as DeleteIcon } from "react-icons/md";
 import { RiFileAddLine as CatalogAddIcon } from "react-icons/ri";
 
 export {
   CatalogAddIcon,
+  CopyIcon,
   DeleteIcon,
   EditIcon,
   GithubIcon,
   GmailIcon,
   LinkIcon,
   LogoutIcon,
-};
+  ShareIcon};
