+++ src/components/shadcn/skeleton.tsx
@@ -1,0 +1,15 @@
+import { cn } from "~/utils/shadcn-helper"
+
+function Skeleton({
+  className,
+  ...props
+}: React.HTMLAttributes<HTMLDivElement>) {
+  return (
+    <div
+      className={cn("animate-pulse rounded-md bg-primary/10", className)}
+      {...props}
+    />
+  )
+}
+
+export { Skeleton }
